requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.12.2
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.22.0
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==4.4.2
docopt==0.6.2
einops==0.8.1
elevenlabs==2.14.0
encodec==0.1.1
ffmpeg-python==0.2.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
g2pkk==0.1.2
grpcio==1.74.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.3.1
Markdown==3.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.46
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.3
Pillow==9.0.1
platformdirs==4.4.0
pooch==1.8.2
praw==7.7.0
prawcore==2.3.0
preshed==3.0.10
proglog==0.1.12
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.55.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.10.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.56.1
trio==0.30.0
trio-websocket==0.12.2
triton==3.4.0
TTS==0.22.0
typeguard==4.4.4
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
update-checker==0.18.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
